refactor(sue): parser
- optimized files and imports of parsers

SUE-2
