--- src/helpers/SUE/index.ts
@@ -1,1 +1,0 @@
-export * from './sueParser';
