+++ src/helpers/index.js
@@ -1,7 +1,6 @@
 export * from './augmentedReality';
 export * from './consul';
 export * from './parser';
-export * from './SUE';
 
 export * from './accessibilityListeners';
 export * from './addressHelper';
